Update des repo + Iservice + serviceimpl

+++ Adopte_A_Pet/src/main/java/fr/formation/inti/Iservices/ICategorieService.java
@@ -6,5 +6,5 @@ import fr.formation.inti.entities.Categorie;
 public interface ICategorieService {
 	void save(Categorie categorie);
 	void update(Categorie categorie);
-	void delete(long id);
+	void delete(Integer id);
 }
