+++ Adopte_A_Pet/src/main/java/fr/formation/inti/Iservices/IDepartementService.java
@@ -1,9 +1,5 @@
 package fr.formation.inti.Iservices;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import fr.formation.inti.entities.Adoption;
-import fr.formation.inti.entities.Centre;
 import fr.formation.inti.entities.Departement;
 
 
@@ -13,7 +9,7 @@ import fr.formation.inti.entities.Departement;
 public interface IDepartementService {
 	void save(Departement departement);
 	void update(Departement departement);
-	void delete(long id);
+	void delete(Integer id);
 	Departement findByDeptid(Integer deptid);
 //find by userID
 //find by adoption ID
