+++ Adopte_A_Pet/src/main/java/fr/formation/inti/dao/IDepartementRepository.java
@@ -2,14 +2,13 @@ package fr.formation.inti.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import fr.formation.inti.entities.Adoption;
 import fr.formation.inti.entities.Departement;
 
 
 
 
 
-public interface IDepartementRepository extends JpaRepository<Departement, Long>{
+public interface IDepartementRepository extends JpaRepository<Departement, Integer>{
 	Departement findByDeptid(Integer deptid);
 //find by userID
 //find by adoption ID
